requirements.txt
2to3==1.0
absl-py==0.8.0
alabaster==0.7.12
anaconda-client==1.2.2
apipkg==1.5
asn1crypto==0.24.0
astor==0.8.0
astroid==1.6.6
astropy==2.0.14
atomicwrites==1.3.0
attr==0.3.1
attrs==19.1.0
Babel==2.7.0
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
backports.weakref==1.0.post1
basis==0.1.0
Beaker==1.5.4
beautifulsoup4==4.8.0
bids-validator==1.2.4
bitarray==1.0.1
blaze==0.10.1
bleach==3.1.0
blis==0.2.4
blivet==0.61.15.74
bokeh==1.3.4
boto==2.49.0
boto3==1.17.40
botocore==1.20.40
Bottleneck==1.2.1
Brlapi==0.6.0
bz2file==0.98
certifi==2019.6.16
cffi==1.6.0
chardet==4.0.0
Click==7.0
cloudpickle==1.2.1
clyent==1.2.1
colorama==0.4.1
configobj==4.7.2
configparser==3.8.1
configshell-fb==1.1.25
contextlib2==0.5.5
coverage==3.6b3
cryptography==1.7.2
cupshelpers==1.0
cycler==0.10.0
cymem==2.0.2
Cython==0.29.13
cytoolz==0.10.0
dask==1.2.2
datashape==0.5.2
dcm2bids==2.1.4
dcmstack==0.7.0
decorator==4.4.0
defusedxml==0.6.0
DenyHosts==2.6
di==0.3
dicom==0.9.9.post1
dill==0.3.0
dipy==0.16.0
distributed==1.28.1
dnspython==1.16.0
docopt==0.6.2
docutils==0.15.2
ecdsa==0.13.3
entrypoints==0.3
enum34==1.0.4
et-xmlfile==1.0.1
ethtool==0.8
execnet==1.7.0
fastcache==1.1.0
firstboot==19.5
Flask==1.1.1
Flask-Cors==3.0.8
freeipa==4.7.2
fros==1.0
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.17.1
futures==3.1.1
gast==0.2.2
gevent==1.4.0
google-pasta==0.1.7
grabbit==0.2.5
gradunwarp==2.0a0
graphviz==0.12
greenlet==0.4.15
grin==1.2.1
grpcio==1.23.0
gssapi==1.2.0
h5py==2.9.0
HeapDict==1.0.0
html5lib==1.0.1
httplib2==0.13.1
idna==2.8
imagesize==1.1.0
importlib-metadata==0.19
INDI-Tools==0.0.6
iniparse==0.4
initial-setup==0.3.9.44
iotop==0.6
ipaclient==4.7.2
ipaddr==2.1.11
ipaddress==1.0.16
ipalib==4.7.2
ipaplatform==4.7.2
ipapython==4.7.2
IPy==0.75
ipykernel==4.10.1
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isc==2.0
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
javapackages==1.0.0
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.1
jmespath==0.9.4
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==5.2.0
jupyter-core==4.5.0
jwcrypto==0.4.2
keepalive==0.5
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kitchen==1.1.1
kiwisolver==1.1.0
kmod==0.1
langtable==0.0.31
lazy-object-proxy==1.4.2
llvmlite==0.29.0
locket==0.2.0
lxml==4.4.1
Magic-file-extensions==0.2
Mako==0.8.1
Markdown==3.1.1
MarkupSafe==1.1.1
mat73==0.52
matplotlib==2.2.5
mccabe==0.6.1
mercurial==5.1
mistune==0.8.4
mock==3.0.5
more-itertools==5.0.0
mpmath==1.1.0
msgpack==0.6.1
msgpack-numpy==0.4.4.3
msgpack-python==0.5.6
multipledispatch==0.6.0
murmurhash==1.0.2
nbconvert==5.6.0
nbformat==4.4.0
neo4j-driver==1.6.3
neomodel==3.3.1
neotime==1.7.4
netaddr==0.7.5
netifaces==0.10.4
networkx==2.2
neurdflib==5.0.0.post1
nibabel==2.5.0
nilearn==0.5.2
nipy==0.4.2
nipype==1.2.1
nitime==0.8.1
niworkflows==0.9.6
nose==1.3.7
notebook==5.7.8
ntplib==0.3.2
num2words==0.5.10
numba==0.45.1
numexpr==2.7.0
numpy==1.16.4
numpydoc==0.9.1
odo==0.5.0
olefile==0.46
openpyxl==2.6.3
packaging==19.1
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.1
partd==1.0.0
Paste==1.7.5.1
pathlib==1.0.1
pathlib2==2.3.4
pathmap==0.1.5
patsy==0.5.1
pbr==5.4.2
pciutils==1.7.3
pcp==4.1
pep8==1.7.1
perf==0.1
pexpect==2.3
pickleshare==0.7.5
Pillow==6.1.0
plac==0.9.6
pluggy==0.12.0
ply==3.4
policycoreutils-default-encoding==0.1
preshed==2.0.1
prometheus-client==0.7.1
prompt-toolkit==1.0.18
protobuf==3.9.1
prov==1.5.3
psutil==2.2.1
psycopg2==2.5.1
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.3.7
pyasn1-modules==0.1.5
pybids==0.9.2
pycairo==1.18.1
PyChart==1.39
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.14
pycrypto==2.6.1
pycups==1.9.63
pycurl==7.19.0
pydicom==1.3.0
pydot==1.4.1
pydotplus==2.0.2
pyflakes==2.1.1
Pygments==2.4.2
pygobject==3.22.0
pygpgme==0.3
pygraphviz==1.5
pyinotify==0.9.4
pykickstart==1.99.66.21
pyliblzma==0.5.3
pylint==1.9.5
pyodbc==4.0.27
PyOpenGL==3.1.0
pyOpenSSL==0.13.1
pyparsing==2.4.6
pyparted==3.9
PyPDF2==1.26.0
PyPubSub==3.3.0
pyrsistent==0.15.4
pysmbc==1.0.13
pytest==4.6.5
pytest-forked==1.0.2
pytest-xdist==1.29.0
python-augeas==0.5.0
python-dateutil==2.8.0
python-dmidecode==3.10.13
python-ldap==3.2.0
python-linux-procfs==0.4.9
python-meh==0.25.3
python-nss==0.16.0
python-xapp==1.0.1
python-yubico==1.2.3
pytz==2016.10
pyudev==0.15
pyusb==1.0.0b1
PyWavelets==1.0.3
pyxattr==0.5.1
pyxdg==0.25
PyYAML==5.1.2
pyzmq==18.1.0
qap==1.0.8
qrcode==5.0.1
qt-backport==0.1.3
QtAwesome==0.6.0
qtconsole==4.5.5
QtPy==1.9.0
rdflib==4.2.2
regex==2019.8.19
reportlab==3.5.23
requests==2.25.1
rope==0.14.0
rope-py3k==0.9.4.post1
rtslib-fb==2.1.69
ruamel.ordereddict==0.4.14
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.1.2
s3cmd==2.2.0
s3transfer==0.3.6
sanlock-python==3.7.3
scandir==1.10.0
scdate==1.10.6
schedutils==0.4
scikit-image==0.14.3
scikit-learn==0.20.4
scipy==1.2.2
seaborn==0.9.0
Send2Trash==1.5.0
seobject==0.1
sepolicy==1.1
setproctitle==1.1.6
setroubleshoot==1.1
simplegeneric==0.8.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
slip==0.4.0
slip.dbus==0.4.0
snowballstemmer==1.9.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.3
spacy==2.1.8
SPARQLWrapper==1.8.4
Sphinx==1.8.5
sphinx-rtd-theme==0.4.3
sphinxcontrib-websupport==1.1.2
spyder==3.3.6
spyder-kernels==0.5.1
SQLAlchemy==1.3.7
srsly==0.1.0
SSSDConfig==1.16.5
statsmodels==0.10.1
subprocess32==3.2.6
svgutils==0.3.1
sympy==1.4
tables==3.5.2
targetcli-fb===2.1.fb49
tblib==1.4.0
Tempita==0.5.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
thinc==7.0.8
toolz==0.10.0
tornado==5.1.1
tqdm==4.35.0
traitlets==4.3.2
traits==5.1.2
typing==3.7.4.1
ujson==1.35
unicodecsv==0.14.1
urlgrabber==3.10
urllib3==1.26.4
urwid==1.1.1
versioneer==0.18
virtualenv==16.7.4
wasabi==0.2.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.6
widgetsnbextension==3.5.1
wrapt==1.11.2
wurlitzer==1.0.3
wxPython==4.0.0
xhtml2pdf==0.2.3
xlrd==1.2.0
XlsxWriter==1.2.0
xlwt==1.3.0
xvfbwrapper==0.2.9
yum-langpacks==0.4.2
yum-metadata-parser==1.1.4
zenmap==6.40
zict==1.0.0
zipp==0.6.0
